Flytt addLegacySerializerOption til v2.
Så den kan brukast i legacy axios kall som skjer der.
